Move devtools changes to other PR
