+++ src/background/devtools/protocol.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) 2021 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { Runtime } from "webextension-polyfill";
+import {
+  liftBackground,
+  registerPort as internalRegisterPort,
+} from "@/background/devtools/internal";
+import type { Target } from "@/types";
+
+export const registerPort = liftBackground(
+  "REGISTER_PORT",
+  (target: Target, port: Runtime.Port) => async () => {
+    internalRegisterPort(target.tabId, port);
+  }
+);
